+++ app/Http/Controllers/MProduct/CategoryController.php
@@ -4,22 +4,29 @@ namespace App\Http\Controllers\MProduct;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Policies\MProduct\CategoryPolicy;
 use App\Http\Requests\MProduct\CreateCategoryRequest;
 use App\Http\Requests\MProduct\UpdateCategoryRequest;
 use App\Repositories\MProduct\ICategoryRepository;
 use App\Services\MProductService;
 use App\Services\MUserService;
 use App\Models\MProduct\Category;
-use Roles;
 
 class CategoryController extends Controller
 {
+    private $categoryPolicy;
     private $categoryRepository;
     private $productService;
     private $userService;
 
-    public function __construct(ICategoryRepository $categoryRepository, MUserService $userService, MProductService $productService)
+    public function __construct(
+        CategoryPolicy $categoryPolicy,
+        ICategoryRepository $categoryRepository, 
+        MUserService $userService, 
+        MProductService $productService
+    )
     {
+        $this->categoryPolicy = $categoryPolicy;
         $this->categoryRepository = $categoryRepository;
         $this->productService = $productService;
         $this->userService = $userService;
@@ -32,16 +39,20 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $views = $this->userService->menusRole();
+        $access = $this->categoryPolicy->access();
 
-        if(!Roles::canView('Kategori', $views))
+        if($access->view != 1)
         {
-            return redirect('dashboard');
+            return redirect('dashboard')->with('fail', 'Tidak memiliki hak untuk lihat kategori');
         }
+        else
+        {
+            $categories = $this->productService->categoriesProducts();
 
-        $categories = $this->productService->categoriesProducts();
+            $views = $this->userService->menusRole();
 
-        return view('mproduct.c_index', compact('categories', 'views'));
+            return view('mproduct.c_index', compact('categories', 'access', 'views'));
+        }
     }
 
     /**
@@ -52,11 +63,22 @@ class CategoryController extends Controller
      */
     public function store(CreateCategoryRequest $request)
     {
-        $this->categoryRepository->store($request);
+        $access = $this->categoryPolicy->access();
 
-        return response()->json([
-            'message' => 'Berhasil tambah kategori'
-        ]);
+        if($access->add != 1)
+        {
+            return response()->json([
+                'status' => 'Fail',
+                'message' => 'Tidak memiliki hak untuk tambah kategori'
+            ]);
+        }
+        else
+        {
+            return response()->json([
+                'status' => 'Success',
+                'message' => 'Berhasil tambah kategori'
+            ]);
+        }
     }
 
     /**
@@ -68,11 +90,24 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {   
-        $this->categoryRepository->update($request, $category);
+        $access = $this->categoryPolicy->access();
+
+        if($access->edit != 1)
+        {
+            return response()->json([
+                'status' => 'Fail',
+                'message' => 'Tidak memiliki hak untuk ubah kategori'
+            ]);
+        }
+        else
+        {
+            $this->categoryRepository->update($request, $category);
         
-        return response()->json([
-            'message' => 'Berhasil ubah kategori'
-        ]);
+            return response()->json([
+                'status' => 'Success',
+                'message' => 'Berhasil ubah kategori'
+            ]);
+        }
     }
 
     /**
@@ -83,8 +118,17 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $this->categoryRepository->destroy($category);
+        $access = $this->categoryPolicy->access();
 
-        return redirect()->back()->with('success', 'Berhasil hapus kategori');
+        if($access->delete != 1)
+        {
+            return redirect('dashboard')->with('fail', 'Tidak memiliki hak untuk hapus kategori');
+        }
+        else
+        {
+            $this->categoryRepository->destroy($category);
+
+            return redirect()->back()->with('success', 'Berhasil hapus kategori');
+        }
     }
 }
